+++ app/helpers/ingest_job_helper.rb
@@ -24,26 +24,28 @@ module IngestJobHelper
 
     parent_uuids = []
     local_parent_and_item_repo_solr_docs_to_update = []
-    index_time_plain = Time.now
-    index_time = index_time_plain.iso8601(3)
+    
+    index_time_plain = Time.current
+    index_time_s = RepoSolrDoc.format_as_solr_s(index_time_plain)
+    index_time_dt = RepoSolrDoc.format_as_solr_dt(index_time_plain)
 
     parent_and_item_repo_docs.each do |doc|
       doc_uuid = doc['uuid']
       parent_uuids << doc_uuid
       local_parent_or_item_repo_solr_doc = RepoSolrDoc.find_or_create_by!(uuid: doc_uuid)
 
-      doc['dateIndexed_s'] = index_time
-      doc['dateIndexed_dt'] = format_as_solr_dt(index_time_plain)
+      doc['dateIndexed_s'] = index_time_s
+      doc['dateIndexed_dt'] = index_time_dt
 
       if local_parent_or_item_repo_solr_doc.first_indexed.nil?
-        doc['firstIndexed_s'] = index_time
-        doc['firstIndexed_dt'] = format_as_solr_dt(index_time_plain)
+        doc['firstIndexed_s'] = index_time_s
+        doc['firstIndexed_dt'] = index_time_dt
         local_parent_and_item_repo_solr_docs_to_update << local_parent_or_item_repo_solr_doc
       else
         first_indexed_plain = local_parent_or_item_repo_solr_doc.first_indexed.to_time
         first_indexed = first_indexed_plain.iso8601(3)
         doc['firstIndexed_s'] = first_indexed
-        doc['firstIndexed_dt'] = format_as_solr_dt(first_indexed_plain)
+        doc['firstIndexed_dt'] = RepoSolrDoc.format_as_solr_dt(first_indexed_plain)
       end
     end
 
@@ -141,12 +143,12 @@ module IngestJobHelper
       end
 
       local_repo_capture_solr_doc = RepoSolrDoc.find_or_create_by!(uuid: uuid)
-      capture_solr_doc['dateIndexed_s'] = index_time
-      capture_solr_doc['dateIndexed_dt'] = format_as_solr_dt(index_time_plain)
+      capture_solr_doc['dateIndexed_s'] = index_time_s
+      capture_solr_doc['dateIndexed_dt'] = index_time_dt
       first_indexed_plain = local_repo_capture_solr_doc&.first_indexed&.to_time || index_time_plain
       first_indexed = first_indexed_plain.iso8601(3)
       capture_solr_doc['firstIndexed_s'] = first_indexed
-      capture_solr_doc['firstIndexed_dt'] = format_as_solr_dt(first_indexed_plain)
+      capture_solr_doc['firstIndexed_dt'] = RepoSolrDoc.format_as_solr_dt(first_indexed_plain)
 
       local_repo_capture_solr_docs_to_update << local_repo_capture_solr_doc if local_repo_capture_solr_doc.first_indexed.nil?
 
@@ -178,17 +180,13 @@ module IngestJobHelper
     repo_solr.delete_unseen_captures_below(ingest_request.uuid, seen_capture_uuids)
 
     # do not update first indexed until we successfully return from commit
-    local_repo_capture_solr_docs_to_update.each { |d| d.update_attributes(first_indexed: index_time) }
-    local_parent_and_item_repo_solr_docs_to_update.each { |d| d.update_attributes(first_indexed: index_time) }
+    local_repo_capture_solr_docs_to_update.each { |d| d.update_attributes(first_indexed: index_time_s) }
+    local_parent_and_item_repo_solr_docs_to_update.each { |d| d.update_attributes(first_indexed: index_time_s) }
 
     Delayed::Worker.logger.info('Done ingesting all captures of Item', uuid: ingest_request.uuid)
   end
 
   def extract_title_from_dublin_core(dublin_core)
     Nokogiri::XML(dublin_core).remove_namespaces!.css('title').text.strip.truncate(250, separator: ' ...')
-  end
-
-  def format_as_solr_dt(datetime)
-    datetime.strftime('%Y-%m-%dT%H:%M:%SZ')
   end
 end