+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added a new endpoint for sending updates for single field values. (DR-2775)
+
 ## [2.0.1] - 2024-01-16
 
 ### Fixed